// lib/models/chat-resource.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Message } from './message';
import { User } from './user';
 /**
 * 
 *
 * @export
 * @interface ChatResource
 */
export interface ChatResource {

    /**
     * @type {User}
     * @memberof ChatResource
     */
    user?: User;

    /**
     * @type {Message}
     * @memberof ChatResource
     */
    lastMessage?: Message;
}